refactor: move the Claude hook endpoint out of index.ts (#548 step 3g)
Every Stop / Notification / Pre|PostToolUse / SessionStart POSTs to /api/hook, and
one request fans out to the session's attention flags, a push to the user's phone,
the tool-call history, the header prompt and the AI title. That fan-out is what
made this long, not the route, so all of it moves to routes/hook-routes.ts.

What index.ts owns arrives as deps: the session lifecycle (setWorking/setWaiting/
publishActivity), the title bookkeeping, the tool stores, the dir-config publish
and the UI port. Each helper is a top-level function taking them explicitly.

Verified against main by booting both and POSTing all seven hook events plus a
non-uuid session id: identical responses, identical activity flags, and an
identical recorded tool-call history.
